+++ featuretools/primitives/transform/geo/latitude.py
@@ -1,0 +1,31 @@
+import numpy as np
+import pandas as pd
+from woodwork.column_schema import ColumnSchema
+from woodwork.logical_types import BooleanNullable, Double, LatLong
+
+from featuretools.primitives.core.transform_primitive import TransformPrimitive
+from featuretools.primitives.utils import _haversine_calculate
+
+class Latitude(TransformPrimitive):
+    """Returns the first tuple value in a list of LatLong tuples.
+       For use with the LatLong logical type.
+
+    Examples:
+        >>> latitude = Latitude()
+        >>> latitude([(42.4, -71.1),
+        ...            (40.0, -122.4),
+        ...            (41.2, -96.75)]).tolist()
+        [42.4, 40.0, 41.2]
+    """
+
+    name = "latitude"
+    input_types = [ColumnSchema(logical_type=LatLong)]
+    return_type = ColumnSchema(semantic_tags={"numeric"})
+    description_template = "the latitude of {}"
+
+    def get_function(self):
+        def latitude(latlong):
+            latlong = np.array(latlong.tolist())
+            return latlong[:, 0]
+
+        return latitude
